consistently pin against Lotus v1.9.0
We were using master since we relied on a patch from Anton which made
"make debug" deals significantly faster. Luckily, that patch is now in
the v1.9.0 release made recently, so we can pin that now.

We make CI follow that, and we update the docs.

Note that Circle manually built and installed binaries, when that's
unnecessary - the integration tests deal with it already.

The fix for our "lotus-miner wait-api" bug is also in master, though
note that it's not in v1.9.0 yet. I did test it with master briefly
though, and it worked like a charm. The code using it is made a bit
better by not spamming unhelpful log lines every second.

Finally, I was running into timeouts starting devnet, since "lotus-miner
init" was downloading the parameter files. One is as big as 1GiB, and
the 60s timeout is too short for my internet connection.

Run "lotus fetch-params" synchronously, showing its progress, which will
work better for those with slower internet. It will also be more visual
than seemingly hanging for minutes without printing anything or using
much CPU.

Fixes #41.